著手建立Boss在BossScene

+++ ShamanKingProject/Assets/Scripts/GrassSnakeScripts/Grasssnake_BT/Grasssnake_ExternalBehaviorTrees/Boss01_EBTs/FirstBossVariables.cs
@@ -38,9 +38,9 @@ public class FirstBossVariables : MonoBehaviour
     [SerializeField, Tooltip("審判之炎爆計時器")]
     private float explosionJudgmentTimer_;
 
-    public float FirstBossHp { get { return firstBossHp_; } set { firstBossHp_ = value; } }
+    public int FirstBossHp { get { return firstBossHp_; } set { firstBossHp_ = value; } }
     [SerializeField, Tooltip("Boss血量")]
-    private float firstBossHp_;
+    private int firstBossHp_;
     public float DistanceFromPlayer { get { return distanceFromPlayer_; } set { distanceFromPlayer_ = value; } }
     [Tooltip("與玩家的距離")]
     private float distanceFromPlayer_;
